A simple math exercise generator added to the project
Two random numbers generated between 1 and 5 and addition or multiplication
chosed. Then the exercise is displayed.

With button B the user can add 1 to the result. With button A the user can
finalize the result. If the result is equal to the true result => happy
face displayed, otherwise sad face displayed:)

Changes to be committed:
	new file:   02_math_exercise/02a_math_exercise.py